// idfistafor/src/app/shared/models/profilio-utente.model.ts
/*!
 * SPDX-FileCopyrightText: (C) Copyright 2023 Regione Piemonte
 * 
 * SPDX-License-Identifier: EUPL-1.2
 */
export enum ProfilioUtente {
    NO_DATA = 0,
    CITTADINO = 1,
    GESTORE = 2,
    UFFICIO_TERRITORIALE = 3,
    COMUNE = 4,
    CONSULTAZIONE = 5,
    SPORTELLO = 7,
    SPORTELLO_GESTORE = 8
}